add webp funcs
